Move bar shape file
